+++ example/lib/screens/theming/typography.dart
@@ -44,7 +44,7 @@ class _TypographyPageState extends State<TypographyPage> {
               onChanged: (c) => setState(() => color = c),
               value: color,
               items: [
-                ComboboxItem(
+                ComboBoxItem(
                   child: Row(children: [
                     buildColorBox(Colors.white),
                     const SizedBox(width: 10.0),
@@ -52,7 +52,7 @@ class _TypographyPageState extends State<TypographyPage> {
                   ]),
                   value: Colors.white,
                 ),
-                ComboboxItem(
+                ComboBoxItem(
                   child: Row(children: [
                     buildColorBox(const Color(0xE4000000)),
                     const SizedBox(width: 10.0),
@@ -62,7 +62,7 @@ class _TypographyPageState extends State<TypographyPage> {
                 ),
                 ...List.generate(Colors.accentColors.length, (index) {
                   final color = Colors.accentColors[index];
-                  return ComboboxItem(
+                  return ComboBoxItem(
                     child: Row(children: [
                       buildColorBox(color),
                       const SizedBox(width: 10.0),